calling script from antoher script

+++ deleteoldfile.sh
@@ -6,10 +6,16 @@ R="\e[31m"
 G="\e[32m"
 N="\e[0m"
 
-if [ -d $filepath/$files ]    # Checks that directory is exists or not.
+if [ -d $filepath ]    # Checks that directory is exists or not.
 then
-    rm -rf $files
-    echo -e "$G $files are deleted successfully $N"
+    if [ -d $filepath/$files ]
+    then
+        rm -rf $files
+        echo -e "$G $files are deleted successfully $N"
+    else
+        echo -e "$R $files are NOT available $N"
+    fi
+
     if [ ! -d $filepath/$files ]
     then
         echo "$files are deleted successfully"
